requirements.txt
dynaconf==3.2.*
pydantic==1.10.*
python_dateutil==2.8.*
urllib3==1.26.*
aenum==3.1.*
requests==2.31.*
snowflake-connector-python[pandas]==3.2.0
snowflake-snowpark-python[pandas]==1.7.0
tenacity==8.2.*
jsonpath-ng==1.5.*
pandas==1.5.*
cryptography==41.0.4
pyopenssl==23.2.0
openpyxl==3.1.2
PyYAML==6.0.1